django/requirements.txt
Django==2.1.4
djangorestframework==3.9.0
psycopg2==2.8.6
django-cors-headers==3.2.1
requests
Pillow